refactor : (#42) add variable
